pyproject.toml
[project]
name = "financetrack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.8.1",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "cryptography==44.0.2",
    "defusedxml==0.7.1",
    "django==5.1.7",
    "django-db-logger==0.1.13",
    "django-mathfilters==1.0.0",
    "django-widget-tweaks==1.5.0",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt==5.5.0",
    "djoser==2.3.1",
    "gunicorn==23.0.0",
    "idna==3.10",
    "oauthlib==3.2.2",
    "packaging==25.0",
    "pillow==11.1.0",
    "psycopg2-binary==2.9.10",
    "pycparser==2.22",
    "pyjwt==2.9.0",
    "python-decouple==3.8",
    "python3-openid==3.2.0",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "six==1.17.0",
    "social-auth-app-django==5.4.3",
    "social-auth-core==4.5.6",
    "sqlparse==0.5.3",
    "urllib3==2.4.0",
]
